Time: 51 ms (58.78%) | Memory: 19.1 MB (18.08%) - LeetSync

+++ koko-eating-bananas/koko-eating-bananas.cpp
@@ -21,7 +21,7 @@ public:
     // O(n)
     inline auto checkK(vector<int>& piles, double k, int h) -> bool {
 
-        long int hours = 0; 
+        long long hours = 0; 
         for (int i = piles.size() - 1; i >= 0; i--){
             double val = ceil(piles[i] / k);
             if (val > h || hours + val > h) return false;
